Added Gestures Samples
